% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.r
\name{visualize}
\alias{visualize}
\title{Building a function in order to replicate 'lineplot' functions in Seaborn
Draw a line plot with possibility of several semantic groupings
The relationship between x and y can be shown for different subsets
of the data using the hue, size, and style parameters
These parameters control what visual semantics are used
to identify the different subsets}
\usage{
visualize(x, y, hue, data)
}
\arguments{
\item{x, }{y Variables that specify positions on the x and y axes}

\item{hue}{Grouping variable that will produce lines with different colors}

\item{data}{Input data structure
Either a long-form collection of vectors
that can be assigned to named variables or
a wide-form dataset that will be internally reshaped}

\item{size}{Grouping variable that will produce lines with different widths}

\item{style}{Grouping variable that will produce lines with different dashes and/or markers}
}
\value{
The matplotlib axes containing the plot
}
\description{
Building a function in order to replicate 'lineplot' functions in Seaborn
Draw a line plot with possibility of several semantic groupings
The relationship between x and y can be shown for different subsets
of the data using the hue, size, and style parameters
These parameters control what visual semantics are used
to identify the different subsets
}
\details{
By default, the plot aggregates over multiple y values
at each value of x and shows an estimate of the central
tendency and a confidence interval for that estimate
}
\seealso{
\href{https://seaborn.pydata.org/generated/seaborn.lineplot.html}{Seaborn}
}
\author{
Eyvaz Najafli
}
